continued work on system messages - renamed object field, implemented custom class for system messages, fixed bold names, implemented created room case
